+++ AlchemistCode/Powers/SlagPower.cs
@@ -1,0 +1,21 @@
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.Entities.Powers;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.Models;
+using MegaCrit.Sts2.Core.ValueProps;
+
+namespace Alchemist.AlchemistCode.Powers;
+
+public class SlagPower : AlchemistPower
+{
+    public override PowerType Type => PowerType.Buff;
+    public override PowerStackType StackType => PowerStackType.Counter;
+
+    public override async Task AfterCardExhausted(PlayerChoiceContext choiceContext, CardModel card, bool causedByEthereal)
+    {
+        if (card.Owner != Owner.Player) return;
+        Flash();
+        await CreatureCmd.GainBlock(Owner, Amount, ValueProp.Move, null);
+    }
+}